Move validator to services folder
