+++ app/component/form-login/index.js
@@ -2,9 +2,9 @@ import {Component} from '../parent-component.js';
 import {FormInput} from '../form-input';
 import {FormHeader} from '../form-header';
 import {FormFooter} from '../form-footer';
-import CredentialValidator from '../../utils/validator.js';
+import CredentialValidator from '../../services/validator';
 import {UserCredentials} from '../../models/user-credentials';
-import {MinLengthRule, NotEmptyRule, RegexpRule} from '../../utils/rules';
+import {MinLengthRule, NotEmptyRule, RegexpRule} from '../../services/validator/rules';
 
 /**
  * Used for defining login input.
